backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
flask-socketio==5.3.6
python-dotenv==1.0.1
alpaca-trade-api==3.2.0
gevent==24.2.1
gevent-websocket==0.10.1
gunicorn==22.0.0
python-socketio==5.7.2
